fix deactivate and activate

+++ routes/api.php
@@ -19,9 +19,6 @@ Route::get('test', function () {
     return "Oksssssssssssssss";
 });
 
-
-
-
 // for candidate
 Route::post('/user/register', 'UserController@register');
 Route::post('/user/login',    'UserController@login');
@@ -47,8 +44,8 @@ Route::middleware('auth:sanctum')->apiResource('profiles',    'ProfileController
 Route::get('/all-users', [AdminController::class,'getAllUsers']);
 Route::get('/all-posts', [PostController::class,'allposts']);
 Route::get('/all-categories', [CategoryController::class,'index']);
-Route::patch('/users/{id}/activate', [AdminController::class, 'activate']);
-Route::delete('/users/{id}/deactivate', [AdminController::class, 'deactivate']);
+Route::middleware('auth:sanctum')->patch('/users/{id}/activate', [AdminController::class, 'activate']);
+Route::middleware('auth:sanctum')->delete('/users/{id}/deactivate', [AdminController::class, 'deactivate']);
 
 
 Route::middleware('auth:sanctum')->put("/admins/post-update", [AdminController::class,'updatePostStatus']);
